Begin major rewrite
